Changed Navbar look and updated upload file activity
Completely changed the look of the nav bar buttons although you can only use the events and homepage button. I also updated the look of the uploadfile activity but ium still having issues when I try to actually upload a file, not sure whats wrong with it. I also added the nav bar to the events page and changed the upload button on the dashboard to a FloatingActionButton
if you think I can make something look better